각 조건에 따라서 잠금해제 될 수 있도록 추가
